\name{ExampleData}
\alias{ExampleData}
\docType{data}
\title{
Example Data
}
\description{
Example data is simulated to show how IPBT work. Two replicates for control and treatment group, respectively are generated. True DE genes are also indicated. 
}
\usage{data(ExampleData)}



\examples{
data(ExampleData)

}

